# topics/open-api-generator-gradle/src/main/docs/api.yaml
# OpenAPI 3 has quite a difference with OpenAPI 2, the way the linter works in the editor will be based on this.
openapi: 3.0.0

# This section will contain metadata information about the API. It will be shown on the editor/UI when we publish this
info:
  title: To-Do Board API
  version: 1.0.0
  description: API specification to support daily to-do list operations

# The list of the servers that we can use to use the API. This will be rendered as a combo box in the editor/UI when we publish it, so the user could pick and use the endpoint they prefer.
servers:
  - url: http://dev-api.todoboard.com/v1
    description: DEV-ENV
  - url: http://api.todoboard.com/v1
    description: PROD-ENV

paths:
  /projects:
    get:
      parameters:
        - in: query
          description: find project by name
          name: name
          schema:
            type: string
      operationId: searchProjects
      tags:
        - Projects
      description: retrieve all available projects in paginated way
      responses:
        200:
          description: successfully retrieved all projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
    post:
      operationId: createProject
      tags:
        - Projects
      description: create new project entry
#      parameters:
#        - $ref: '#/components/parameters/X-Request-ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        201:
          description: successfully created a project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequest'


  /projects/{projectId}:
    get:
      operationId: getProjects
      tags:
        - Projects
      description: retrieve a project by id
      parameters:
        - name: projectId
          in: path
          schema:
            type: integer
            format: int64
            minimum: 1
          required: true
          allowEmptyValue: false
          example: 1

      responses:
        200:
          description: the specific project successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'

components:
  schemas:
    ProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 10
          description: the name of the project
          example: To-do Application Development
        additionalName:
          type: string
          description: the name of the project
          example: To-do Application Development
        email:
          type: string
          format: email
        amount:
          type: integer
          minimum: 1
          maximum: 20
        date:
          type: string
          format: date
        password:
          type: string
          format: password
        uuid:
          type: string
          format: uuid
        testField1:
          type: string
        testField2:
          type: string
      required:
        - name
    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: unique id of the project for the identifier
          example: 1
        name:
          type: string
          description: the name of the project
          example: To-do Application Development
  parameters:
    X-Request-ID:
      name: X-Request-ID-name
      in: header
      description: X-Request-ID description
      required: true
      schema:
        type: string
      example: "Example X-Request-ID"